+++ CHANGELOG.md
@@ -7,14 +7,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
- 
+
 ### Added
 
- - Asyncio profiling support
+ - asyncio profiling support
+- `phanos.publisher.NamedLoggerHandler` designed to be used of configuration `profile.dict_config`
 
 ### Changed
+- `job` argument is now mandatory in `config` and `dict_config`, because Flask dependency was removed
+- handlers moved from `phanos.publisher` to `phanos.handlers`
+- `current_node` of `ContextTree` moved into `Contextvar` due to asyncio support
+- limit `requirements` to minimum and separate development ones into `requirements-dev.txt`
+- `messaging` is now defined in this project and thus separated from `imp-prof`
 
- - current_node of MethodTree stored in Contextvar instead of MethodTree because of async
+
+## [0.1.0] - 2023-08-02
+
+
+### Added
+
+- support of dictionary configuration of profiler
 
 
 ## [0.0.0] - 2023-06-01